Plantilla formulario competiciones

+++ src/Easanles/AtletismoBundle/Controller/CompeticionController.php
@@ -14,7 +14,7 @@ class CompeticionController extends Controller
     public function listadoCompeticionesAction() {
     	$repository = $this->getDoctrine()->getRepository('EasanlesAtletismoBundle:Competicion');
     	$competiciones = $repository->findAll();
-        return $this->render('EasanlesAtletismoBundle:Competicion:listado_competiciones.html.twig',
+        return $this->render('EasanlesAtletismoBundle:Competicion:list_competicion.html.twig',
            array('competiciones' => $competiciones));
     }
     
@@ -28,11 +28,15 @@ class CompeticionController extends Controller
     	 	// guardar la tarea en la base de datos
     	 	$em = $this->getDoctrine()->getManager();
     	 	$em->persist($com);
-    	 	$em->flush();
+    	   try {
+    	      $em->flush();
+    	   } catch (\Exception $e) {
+    	   	return new Response($e->getMessage());
+    	   }
     	 	return $this->redirect($this->generateUrl('listado_competiciones'));
     	 }
     	 
-       return $this->render('EasanlesAtletismoBundle:Competicion:formulario_competiciones.html.twig',
+       return $this->render('EasanlesAtletismoBundle:Competicion:form_competicion.html.twig',
              array('form' => $form->createView()));
     }
    
@@ -43,36 +47,60 @@ class CompeticionController extends Controller
     	 $em = $this->getDoctrine()->getManager();
     	 $repository = $this->getDoctrine()->getRepository('EasanlesAtletismoBundle:Competicion');
     	 $com = $repository->findOneBy(array('nombre' => $nombre, 'temp' => $temp));
-    	 $em->remove($com);
-    	 $em->flush();
-    	 return $this->redirect($this->generateUrl('listado_competiciones'));
+    	 if ($com != null){
+    	    $em->remove($com);
+    	    try {
+    	   	$em->flush();
+    	   } catch (\Exception $e) {
+    	   	return new Response($e->getMessage());
+    	   }
+    	    return $this->redirect($this->generateUrl('listado_competiciones'));
+    	 } else {
+       	$response = new Response('No existe la competicion "'.$nombre.'" en la temporada '.$temp.' <a href=".">Volver</a>');
+       	$response->headers->set('Refresh', '2; url=.');
+       	return $response;
+       }
     }
     
     public function editarCompeticionAction(Request $request, $nombre, $temp){
     	$em = $this->getDoctrine()->getManager();
     	$repository = $this->getDoctrine()->getRepository('EasanlesAtletismoBundle:Competicion');
     	$com = $repository->findOneBy(array('nombre' => $nombre, 'temp' => $temp));
-    	 
-    	$form = $this->createForm(new ComType(), $com);
+
+    	if ($com != null) {
+    	   $form = $this->createForm(new ComType(), $com);
     	
-    	$form->handleRequest($request);
+    	   $form->handleRequest($request);
     	
-    	if ($form->isValid()) {
-    		// guardar la tarea en la base de datos
-    		$em->flush();
-    		return $this->redirect($this->generateUrl('listado_competiciones'));
-    	}
+    	   if ($form->isValid()) {
+    	   	try {
+    	   		$em->flush();
+    	   	} catch (\Exception $e) {
+    	   		return new Response($e->getMessage());
+    	   	}
+    	   	return $this->redirect($this->generateUrl('listado_competiciones'));
+       	}
     	
-    	return $this->render('EasanlesAtletismoBundle:Competicion:formulario_competiciones.html.twig',
-    			array('form' => $form->createView()));
-    	
+      	return $this->render('EasanlesAtletismoBundle:Competicion:form_competicion.html.twig',
+    		   	array('form' => $form->createView()));
+       } else {
+       	$response = new Response('No existe la competicion "'.$nombre.'" en la temporada '.$temp.' <a href=".">Volver</a>');
+       	$response->headers->set('Refresh', '2; url=.');
+       	return $response;
+       }
     }
     
     public function verCompeticionAction($nombre, $temp){
-    	$repository = $this->getDoctrine()->getRepository('EasanlesAtletismoBundle:Competicion');
-    	$com = $repository->findOneBy(array('nombre' => $nombre, 'temp' => $temp));
-        	     	 
-    	return $this->render('EasanlesAtletismoBundle:Competicion:ver_competicion.html.twig',
-    			array('com' => $com));
+    	 $repository = $this->getDoctrine()->getRepository('EasanlesAtletismoBundle:Competicion');
+    	 $com = $repository->findOneBy(array('nombre' => $nombre, 'temp' => $temp));
+
+       if ($com != null) {
+          return $this->render('EasanlesAtletismoBundle:Competicion:ver_competicion.html.twig',
+    	          array('com' => $com));
+    	 } else {
+    	 	$response = new Response('No existe la competicion "'.$nombre.'" en la temporada '.$temp.' <a href=".">Volver</a>');
+    	 	$response->headers->set('Refresh', '2; url=.');
+    	 	return $response;
+    	 }
     }
 }
